requirements.txt
bs4==0.0.2
matplotlib==3.9.4
numpy==2.0.2
pandas==2.2.3
polars==1.27.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
requests-ip-rotator==1.0.14
SQLAlchemy==2.0.40
urllib3==1.26.20
streamlit==1.44.1
mplsoccer==1.4.0
